Moving psbchsim to be only UE dependent
Also, this commit includes removing the NR_txUE_PSBCH
struct members into the NR_UE_PSBCH struct. At this point,
I cannot see a reason for a seprate struct. Especially
since the NR_txUE_PSBCH is not declared within the PHY_VARS_UE
scope so it is not accessible by the UE PHY layer global
anyways.

Additionally, the changes in psbchsim will allow the
UE to be used instead of the gNB. It configures the
frame parameters and should allocate memory for the
tx and rx buffers used in SL.