HACKY change for NR_POLAR_PSBCH_E.
This is going to be a long explaination. The spec says
for polar encoding that N = 512 (input size) - 2^9,
K = 56, and E (encoded size) = 1782. However, the interleaver
will break 512 into 16, 32 bit sections. These 16 sections
are moved around to form the interleaved input. Then
the code does something else and basically, every 16th
bit is modified (or important) and triggers the mingroupsize
to be 16. The mingroupsize is calculated in a for loop and
starts from 0->16 and then resets again at 0. However, the
E size if 1782, which is not divisible by 16, will make
minigroupsize end at 6 (it is incremented until the special
character is found at index 16). The hack is to just
make 1782 large enough so the last bit is found, so we change
it to size 1792.