Updated the write_time_stamp value.
Previously this value was based on the slot_tx,
however the slot_tx is in the future and has not
be fully populated at this time (execution of writing
to USRP). So, we should write to USRP in current
slot, but we will fill the buffer in the generation
code in the future.