The nb_prefix_samples0 is dependent on sync status.
By changing the global value of the nb_prefix_samples0
at the top of the thread, we are never able to properly
set the value after sync. To fix this, a local variable
was added in every place we use nb_prefix_samples0 in
the UE_thread_SL. However this commit has quite a few
changes becasue some places where the nb_prefix_samples0
global was used did not have the UE global available, so
we had to update the function footprint to pass in a
sync bool for proeprly setting the local nb_prefix_samples0
variable.