CMDLINE_PARAMS_CHECK_DESC must be correct
When merging in the new --nfapi str CL option from
develop, the santizer was catching a null pointer in
config_checkstr_assign_integer(). To fix this, I logged
params and i in config_execcheck(). This showed that
only 36 CMDLINE_PARAMS_CHECKs were happening when we
have 41 CL options in executables/softmodem-common.h.

To fix this, I added 7 more { .s5 = { NULL } }, for the
remaining CL parameters defined. Furthermore, the .s3a
CL parameter must be associated with the new --nfapi
string, so if it does not properly line up with the CL
definitions order, it will fail. The logs added in
config_execcheck() will show this.