This commit is a large refactoring of UE_thread_sl
We chose to initially create our own (SL) version
of the UE_thread() function in the nr_ue.c. The original
function was then refactored and merged to the develop
branch. Now that it has been refactored, we have added
the UE SL support into the original function and removed
the UE_thread_sl function. This still needs to be validated
on the USRPs prior to merge. (It has been tested with RFSIM).

Also, a few more lingering warnings related to the int32_t -> c16_t
conversion