We are currently not properly freeing our allocations
In this commit, I am removing the calls to the
dlsch and slsch free functions. The only reason
for this, is to clean up the terminal output.
Currently the output will report all the memory
leaks and it is A LOT. This spams the termianl and can
be difficult to check the state of the development
code. We need to ultimately fix this, but for now,
the code will crash during the free - which doesnt
spam the terminal.