This commit validates OFDM modulation and demoduation.
Specifically, we call the nr_ue_pssch_common_procedures(),
which conducts the modulation, and then the gNB common functions
for the rotation and demoduation (nr_slot_fep_ul and
apply_nr_rotation_ul). The demodulation functions were
not changed, and the modulation function was cleaned up,
and modified to take in the num_symbols set in the harq config
rather than being hard coded to 14.