Enabling sample adjustment based on RSRP
This commit will enalbe the UE sync adjustment procedure.
For every slot, it will take the DFT, conduct channel estimation,
calculate RSRP, assume a PSBCH was recieved (since we cannot
guarentee the slot alignment when NOT in RFSIM) and will
adjust the sync position if the RSRP is better than the initial/
previously saved SSB RSRP. This procedure has not effected
RFSIM functionality.