Trigger RX processing after scheduler has been triggered
The scheduler prepares PRACH/PUSCH/PUCCH FAPI PDUs. As of this commit,
the scheduler runs in the TX processing chain, whereas these PDUs are
destined for RX processing.

To avoid data races, instead of triggering the RX processing at the same
time as TX processing in the RU thread, this commit changes the code to
trigger RX processing after the scheduler has been run in TX processing
to avoid data races.