Proposal for power control for rfsimulator
This implements power control in rfSimulator by using 0dBFS power reference.
In RX, 0dBFS power reference is the power recieved in dBm corresponding to
a RMS of a full scale sine wave received on a single subcarrier. That
reference is later used to calculate received power.

For TX, 0dBFS power reference is the power transmitted by the rf device if a
single carrier full scale sin wave is sent to the rf device.

rfSimulator uses a difference between these two values to calculate "digital
sample gain" which is applied to correctly estimate the received power regardless
of rfSim mode.

In order to avoid realtime effect of this change one must configure the
system to run with equal TX/RX power reference. To do that tx_amp_backoff_dB has to
match ssPBCH_BlockPower in your config file.