replace assert() by DevAssert() and AssertFatal()
Because this is how we do it in openair.

I removed some stuff in openair1/SCHED/phy_procedures_lte_eNb.c
that was not added by the new threadpool merge request. That does
not hurt.

Also, assert() is traditionally (as far as I know) used for when the
code is compiled in "debug" mode, and does not emit any check in
"production" mode. But in openair, we see stuff like:

  int *x = malloc(sizeof(int));
  assert(x != NULL);

Which is (to me) a clear abuse of assert() because we want to check
malloc() (and others) all the time, and react appropriately.

So let's get rid of this assert(), as it seems people don't really
understand what's the use case of this macro.