rework threadpool management
This work was done by Mikel Irazabal.

The original branch was too outdated to rebase properly on top of
develop, plus the commit messages were not informative. Also, the
code was adapted a bit to fit the project (no use of assert(),
clang formatting, change some names which were not easy to
understand).