nr pdcp benchmarking: adapt to run on top of current develop
Lots of changes were done, so there was first a merge of the develop branch
into nr-pdcp-benchmarking.

Then this commit adapts the benchmark to the new NR PDCP security code,
including some hacks to compile it.

The previous version of the benchmarking code is available in the commit
0033f2840e1426db87ea4004e2d6af723e054e62.

On my test machine, this new version (using openssl and not nettle anymore)
is slower. To be investigated (and fixed, if possible).

(This work is not to be merged into develop. This branch must not be
deleted without asking me or Raymond.)