UE Send only TX symbols in mixed slot
So far we send the full slot to radio in case of mixed slot. This might
work if separate TX and RX ports of the radio are used but still is not
good becuase the PA will disturb the DL symbols. When only single port
on the radio is used, then we will not receive DL symbol in the mixed
slot becuase TX path is chosen for the entire slot.

In this commit, we start to transmit at the guard symbol of the mixed
slot so we receive the DL symbols and have enough time for PA warmup.