Add error message in sync req handler
Sync request from MAC should either have SSB position that the PHY has
to sync with or instruct the PHY to do a scan. If neither is provided,
an error message is displayed and PHY proceeds to perform sync with
default command line SSB position.

In SA mode, the error message is valid indicating something went wrong
in the MAC. In do-ra/phytest mode, the RRC config loaded from file could
have a configuration that sends reconfig with sync during
initialization. This causes the error message to be displayed but can be
ignored as it is irrelevant.