Correct FFT error in DC bin
There is a loss in precision when (2^15)-1 integer MAX value is
multiplied and right shifted by 15. This happens to the first element in
a butterfly thus increasing error only in this bin as it is a summation
of all inputs. Instead, we use a 14 bit twiddle factor so the
multiplication will preserve the original sample value without any loss.
Simulation results show that this reduces the error in DC by more than
90%.

Add 14 bit twiddle for 16, 64, 256, 1024 and 2048 DFT.

The MR should have the simulation results attached.