Refactor XML test execution in CI and introduce common test cases
The goal of this refactoring is to simplify the XML test execution by removing TestCaseRequestedList
and adopt an approach based on importing common test cases with XInclude. The common test cases
are listed in the new file common.xml and can be reused by different CI tests as subsequent actions.
To add a subsequent action after a main test case is executed, "postActions" tags are introduced.

The main.py is also refactored following this approach: (1) extend the OaiCiTest by adding all
fields configurable by xml (2) introduce a function LoadTests to load all test cases from
an xml file and store them in a dictionary (3) 2 dictionaries are present: (a) the main test
cases (b) the common test cases (4) the main script will loop through all tests in the main
dictionary and if a postAcion is present it fecthes it from the common dictionary and executes
the test

The `ExecuteActionWithParam` function has been updated to remove direct access to XML elements
and instead uses attributes directly from the `test` object

A `do_test` function was introduced to centralize the logic for executing a test case.

Removed outdated checks for test exclusions and requested tests, streamlining the process by
directly using the provided test list in the XML. Default values were added instead.