Return 0 size in seq_arr_size instead of asserting
The returned size is generally used to loop through
the seq_arr, in the current code, it is used in 2 scenarios:
(1) for loop (2) qsort. Both these iterations will be skipped
if size is 0, so seq_arr NULL.

seq_arr: Return 0 in seq_arr_size() when input is NULL

Instead of asserting, return 0 when the input `seq_arr_t*` is NULL.
This improves robustness in calling code, where `seq_arr_size()` is
used in safe contexts like `for` loops and `qsort`, both of which
gracefully handle a size of 0.

This change avoids unnecessary crashes due to NULL pointer access
in non-critical paths, e.g. in the upcoming commits, when looping
through the PDU Session setup list, before the list is initialised.