"find" library helpers: make inputs const, return {found=false} on failure
The find lib functions are returing an elm_arr_t type struct
which is containing a "found" bool member, which can be use to
signal externally that no element was found. This avoid asserting
on the seq_arr being equal to NULL, which is already checked by
seq_arr lib functions. This change is necessary since the find
lib will be adopted in PDU Session handling and it may happens that
a find function is called when the PDU Session seq_arr is not
initialized yet, leading to an assert.

This commit:

* Mark read input parameters to find helpers (find_if, find_if_arr_it) as const
* Ensure failures return a {found = false, it = NULL} element