PDSCH compensate channel amplitude
The PDSCH channel compensation function compensates only phase and not
the magnitude. Compensating the phase is not necessasary because the
channel magnitude of each RE is used as LLR thresholds. Hence, the LLRs
that are close to origin have the same likelihood as when received
samples are compensated for magnitude.

However, compensating magnitude shows a clean constellation and LLR
plot. So this commit adds it as a optional feature.