WIP: fix(oran_fhlib_5g): Use new RX PRB mapping
K release introduced a new structure to describe the segmentation of RX symbols.
This commit enables using this structure to retrieve RX data.

The commit is incomplete.
As of this commit, the PRACH buffer is initialized to 0 in order to avoid a segmentation fault but it should normally be initialized using `oran_allocate_cplane_buffers`.