NR UE: actually connect, read and write to the multiple RF frontends in parallel
When reading samples, we now read from all RF frontends but only store the data from the one where the UE is currently connected to.
Also when writing samples, we now write to all RF frontends, but the actual data only to the one where the UE is currently connected to, zeroes to the others.
When setting a frequency, we switch the currently connected RF frontend based on the distance to the initially specified cell frequencies.