RRC PDU Session Release: refactor rrc_gNB_process_NGAP_PDUSESSION_RELEASE_COMMAND
3GPP TS 38.413 8.2.2 PDU Session Resource Release says:

> upon reception the NG-RAN node shall execute the release of the requested PDU sessions.
> For each PDU session to be released the NG-RAN node shall release the corresponding
> resources over Uu and over NG, if any.

Therefore, any PDU session in the pdusession_release_params list that is
also present in the RRC list can be set to release, regardless of its status.

The rrc_gNB_send_NGAP_PDUSESSION_RELEASE_RESPONSE is sent upon completion
of the RRCReconfiguration and needs to contain released PDU Sessions (mandatory IE),
therefore it is no necessary to send it in case there is no PDU session to release.

The function signature was also updated along with other minor changes.